📝  Fix code block format in comparison table (#28)
* docs: Update FAQ

Fix indentation, grammar and some wording. Correct some factual claims or make them more precise (e.g. engine vs. connection string)

* 🔧 fix tralling issue

* docs: Update claims that have changed/will change in the future

SQLModel just published a new release, so I updated the claim about it not having done so in a while. I also added a little notice about when I called ormdantic brand new, unproven and exciting, as at least the first aspect will change in the future.

* docs(faq): Convert comparison to table

This seems to be more concise. Unfortunately, the code blocks had to be simplified a bit, but they should still be helpful.

* docs(faq): Add tgn file for the comparison table

This way the table can be easily edited on tablesgenerator.com instead of an editor, which would be error-prone..

* Revert "docs(faq): Add tgn file for the comparison table"

This reverts commit 8fcec5f88a8abc4b45d76cea0c1a295e961f5663.

* docs(faq): Fix code block format 

Unfortunately, the code blocks in the comparison table are displayed differently in mkdocs than they are in my Markdown preview. This commit should fix it.

* docs(faq): Remove unnecessary html tags

After building the site and serving it, I discovered that the previous commit introduced some unnecessary html tags, so I removed them.

Co-authored-by: Yasser Tahiri <yasserth19@gmail.com>