fix: Improve field type translation
Every subclass of the listed basic types is assumed to be meant to be translated to the corresponding SQLAlchemy types.  Adds support for the Boolean, Date and DateTime types, fixes an issue where e.g. ConstrainedInt would be translated to JSON type.

Since this changes the database type for some python types, it might break code or lead to unexpected results when regenerating tables that were first created before this change.